% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbasingle.R
\name{fbasingle_ui}
\alias{fbasingle_ui}
\title{fbasingle_ui}
\usage{
fbasingle_ui(title = "")
}
\arguments{
\item{title}{a text}
}
\value{
shiny tag list
}
\description{
fbasingle_ui
}
\author{
Reinhard Simon
}
