% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_analysis.R
\name{fieldbook_analysis}
\alias{fieldbook_analysis}
\title{fieldbook_analysis}
\usage{
fieldbook_analysis(input, output, session, values)
}
\arguments{
\item{input}{shiny}

\item{output}{shiny}

\item{session}{shiny}

\item{values}{shiny}
}
\description{
fieldbook_analysis
}
\author{
Reinhard Simon
}

