% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_analysisAddin.R
\name{fieldbook_analysisAddin}
\alias{fieldbook_analysisAddin}
\title{fieldbook_analysisAddin}
\usage{
fieldbook_analysisAddin(fieldbook = NULL)
}
\arguments{
\item{fieldbook}{a dataframe}
}
\value{
string message
}
\description{
Adds an addin to the addins menu and shows a view within RStudio.
}
\seealso{
Other addin: \code{\link{locationsAddin}}
}
