% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_ui.R
\name{locations_ui}
\alias{locations_ui}
\title{locations_ui}
\usage{
locations_ui(title = "Locations UI")
}
\arguments{
\item{title}{string}
}
\value{
shiny taglist
}
\description{
locations_ui
}
\author{
Reinhard Simon
}

