% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsAddin.R
\name{locationsAddin}
\alias{locationsAddin}
\title{locationsAddin}
\usage{
locationsAddin()
}
\value{
string status
}
\description{
Displays a dialog view in RStudio and is accessible via the Addins menu.
}
\seealso{
Other addin: \code{\link{fieldbook_analysisAddin}}
}
